suggested to use MIT license and add <wasd> to movement
As there is no license, the software is hard to be used.  Hence suggest to add a license.  At the same time, many games use wasd for cursor movement and suggest that as well.

There are two issues outstanding: (a) with the setxxx placement but that is only style warning and (b) more problematic is sometimes the main thread does not release the display function and the system hang.